:hammer: Refactor txs schema
After reading the docs, it turns out that the abci::Log and abci::Info
fields can be arbitrary data and are not gauranteed to be JSON or an
error message/etc.

Instead of having logic to try and capture errors messages, store the
log string and info string without attempting conversion. Then store the
actual events from the tendermint-abci library response.